Add the (slightly broken) menu value updater scripts
