--- Assets/Universal/Scripts/PauseMenu/FovValueUpdater.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5bf08352c6bf827408ae80bbb3068935
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
