--- Assets/Universal/Scripts/PauseMenu/MouseSensValueUpdater.cs
@@ -1,11 +1,0 @@
-using UnityEngine;
-using TMPro;
-public class MouseSensValueUpdater : MonoBehaviour
-{
-    public TextMeshProUGUI MouseSensText;
-    
-    public void Update()
-    {
-
-    }
-}
