--- Assets/Universal/Scripts/PauseMenu/MouseSensValueUpdater.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d30d21186f1916046bed97d3fc0041f1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
