+++ Assets/Universal/Scripts/PauseMenu/ValueUpdater.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f0291254007524f41aae59eaa3e44c6d
+guid: 8c0a6422f52eaf846b1da4b560fc42f3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
